Users RUD
Added admin role, added user list and user edit form, User RUD logic, admin granted for fist user, second user added without role

+++ MyBasicTaskManager/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
-﻿using MyBasicTaskManager.Models;
+﻿using AutoMapper;
+using MyBasicTaskManager.Models;
 using MyBasicTaskManager.Services;
 using System;
 using System.Collections.Generic;
@@ -8,7 +9,7 @@ using System.Web.Mvc;
 
 namespace MyBasicTaskManager.Controllers
 {
-    [Authorize]
+    [Authorize(Roles ="admin")]
     public class UsersController : Controller
     {
         private readonly UsersService _usersService = new UsersService();
@@ -16,29 +17,35 @@ namespace MyBasicTaskManager.Controllers
         public ActionResult Index()
         {
             ViewBag.Title = "Manage Users";
-            var model = new UsersViewModel();
+            var model = new UsersViewModel()
+            {
+                Users = _usersService.GetAll(),
+                CurrentUserId = _usersService.GetCurrentUserId()
+            };
             return View(model);
         }
-        public ActionResult UserForm(bool IsExisting, int Id = 0)
+        public ActionResult UserForm(bool IsExisting, string Id)
         {
             var Viewmodel = new UserFormViewModel();
+            var config = new MapperConfiguration(cfg => {
+                cfg.CreateMap<UserFull, UserFullViewModel>();
+            });
+            IMapper mapper = config.CreateMapper();
             if (IsExisting)
             {
-                var model = _usersService.Get(Id);
-                ViewBag.Title = "Edit task";
+                ViewBag.Title = "Edit user";
                 Viewmodel.IsExisting = true;
-            }
-            else
-            {
-                ViewBag.Title = "Add new task";
-                Viewmodel.IsExisting = false;
+                var user = _usersService.Get(Id);
+                Viewmodel.User= mapper.Map<UserFull, UserFullViewModel>(user);
+                if (user.Roles.Count() > 0)
+                    Viewmodel.User.IsAdmin = true;
             }
             return View(Viewmodel);
         }
         [HttpPost]
         public ActionResult UserForm(UserFormViewModel userFormViewModel)
         {
-            //userFormViewModel.User.TryUpdateModel();
+            TryValidateModel(userFormViewModel.User);
             if (ModelState.IsValid)
             {
                 _usersService.Save(userFormViewModel.IsExisting, userFormViewModel.User);
@@ -49,9 +56,10 @@ namespace MyBasicTaskManager.Controllers
                 return View(userFormViewModel);
             }
         }
-        public ActionResult DeleteUser(int Id)
+        public ActionResult DeleteUser(string Id)
         {
-            _usersService.Delete(Id);
+            if(Id!=_usersService.GetCurrentUserId())
+                _usersService.Delete(Id);
             return RedirectToAction("Index");
         }
     }
